+++ src/library-authoring/add-content/PickLibraryContentModal.test.tsx
@@ -9,7 +9,7 @@ import {
 import { studioHomeMock } from '../../studio-home/__mocks__';
 import { getStudioHomeApiUrl } from '../../studio-home/data/api';
 import mockResult from '../__mocks__/library-search.json';
-import { LibraryProvider } from '../common/context';
+import { LibraryProvider } from '../common/context/LibraryContext';
 import { ComponentPicker } from '../component-picker';
 import * as api from '../data/api';
 import {
