+++ src/library-authoring/component-info/ManageCollections.test.tsx
@@ -12,7 +12,7 @@ import mockCollectionsResults from '../__mocks__/collection-search.json';
 import { mockContentSearchConfig } from '../../search-manager/data/api.mock';
 import { mockContentLibrary, mockLibraryBlockMetadata } from '../data/api.mocks';
 import ManageCollections from './ManageCollections';
-import { LibraryProvider } from '../common/context';
+import { LibraryProvider } from '../common/context/LibraryContext';
 import { getLibraryBlockCollectionsUrl } from '../data/api';
 
 let axiosMock: MockAdapter;
