+++ src/library-authoring/components/ComponentCard.test.tsx
@@ -5,7 +5,7 @@ import {
   waitFor,
   initializeMocks,
 } from '../../testUtils';
-import { LibraryProvider } from '../common/context';
+import { LibraryProvider } from '../common/context/LibraryContext';
 import { getClipboardUrl } from '../../generic/data/api';
 import { ContentHit } from '../../search-manager';
 import ComponentCard from './ComponentCard';
